[refactor] Altered the DSL so that Trait Collections are not bound to models (30m)
Before, SetBuilder provided a DSL for declaring traits on a model and then required you to pass that model to a Set.

This forced the client to store collections of traits in instance variables on classes — i.e. as globals.

In order to allow clients to store trait collections in local variables and pass them in to sets (to avoid all the pitfalls of global variables), we've refactored the DSL so that you contsruct Trait Collections independently of models and pass them into sets.